+++ core/src/com/unibo/s3/main_system/communication/NamedActor.scala
@@ -1,4 +1,4 @@
-package daniele.comunication
+package com.unibo.s3.main_system.communication
 
 import akka.actor.UntypedAbstractActor
